pyproject.toml
[tool.poetry]
name = "helpers"
version = "0.1.0"
description = ""
authors = ["alexey_naidiuk <alexeynaidiuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = {extras = ["all"], version = "^0.27.2"}
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
